requirements.txt
celery==5.4.0
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==23.0.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
SQLAlchemy==2.0.34
validators==0.34.0
